[hdSt] Delete testHdStIndirectDrawBatchCodeGen.
This test specifically checked the text of the diagnostic output when we failed a shader compile, which isn't really an appropriate thing to test and it wasn't clear how to rehabilitate it into something useful.

For failed shader compiles, we think that testUsdImagingGLInvalidMaterial should give us the coverage we need.

(Internal change: 2378454)