Fix enable scene materials in usdview
The PrimTypePruningSceneIndex was using the HdMaterialBindings
Schema to determine what materials to prune, however this is
run before the materialResolvingSceneIndex where those bindings
are added.
This change replaces the PrimTypePruningSceneIndex for the
SceneMaterialPruningSceneIndex, and inserts it later in the
scene index chain so that it is constructed after the the cards
mesh/geomsubsets/materials are added. Cards bindings are
not dirtied because they are already marked as final, and to
make sure the cards materials themselves are not pruned we
apply the HdBuiltinMaterialSchema.

Fixes #3573

(Internal change: 2378485)
(Internal change: 2378503)