vdf: specialize keep mask accumulation for speculation nodes
Determining keep masks is typically a small portion of overall
scheduling costs.  However, speculation vectorization can produce a
pathological input for the previous implementation when a read/write
output, which may pass its buffer, has many speculations that read
from it.  Without vectorization, there are n connections to n
speculation nodes, each with a single input and output, so this scales
linearly in the number of speculations.

However, we vectorize speculations together into a single node with n
inputs and n outputs to amortize the overhead of internal speculation
scheduling and evaluation.  Unfortunately, in the scenario above, this
causes _SetBufferPassDataForOutputs to visit n outputs for each of the
n input connections.  Additionally, though VdfSpeculationNode's
_ComputeInputDependencyMask method takes advantage of the knowledge
that speculation inputs can only depend on the output with the same
name, VdfNode::GetOutput is implemented with linear search.  This
yields a worst-case O(n^3) time complexity.  As a practical result, we
see a vectorized speculation node with ~2400 input/output pairs
driving ~3 seconds of overhead in _SetBufferPassDataForOutputs.

While we could specialize the input connection dependency query as it
is currently stated by the scheduler, instead, consider that the role
of speculation nodes in the system is to permit evaluation in the
presence of topological cycles that are not sparse data-dependency
cycles.  As such, speculation nodes do not actually read values across
their input connections; they use an internal executor to invoke
whichever nodes are needed to produce the speculation outputs.  So,
the result of _ComputeInputDependencyMask does not necessarily capture
the potentially relevant keep bits for a speculation node anyway.
Therefore, we skip speculation node input dependencies entirely for
the purpose of determining the keep mask.

(Internal change: 2378936)