Remove Get/Set UseSceneMaterials from HdStRenderPassState since this toggling is handled by the SceneMaterialsPruningSceneIndex
Since the materialIsFinal is now being handled by scene indices we remove the now unnecessary check in Storm's drawBatch.

In UsdImaging 1.0 the material adapter returned an empty
VtValue for the material when scene materials were not
enabled, so this check would not really do anything for regular
materials, and result in incorrect behavior for cards because
it would switch to the fallback material.

Fixes #3573

(Internal change: 2378975)