sdf: Defer node creation for lists and functions in variable expressions
Previously, the variable expression parser would create expression
tree nodes for items in a list or arguments to a function immediately
after parsing those elements. This change makes it so that these nodes
are created only when the node for the list/function is requested
(i.e., after all elements have been parsed). This opens the door for
the parsing code to produce different types of nodes in the future.

There should be no change in behavior, except that in certain cases
error messages will no longer include position information. There was
already a pre-existing inconsistency where some error messages included
this info and some didn't, so if we need to address this we should do
so as a separate cleanup.

(Internal change: 2379141)
(Internal change: 2379399)