Update MaterialX version checks
This adds a version calculation that combines the major/minor/build versions
into one value for easier version comparisons.

Similar functionality was added to MaterialX in v1.39.4 however we would
need to wait until that version is the minimum version to be able to use it.

This change was inspired by PR #3653 by @scotbrew

Fixes #3651
Closes #3653

(Internal change: 2379338)
(Internal change: 2379370)