[hdPrman] HdPrman now supports specifing pixel filter parameters (Ri:PixelFilterName and Ri:PixelFilterWidth) in scene-provided render settings such as UsdRenderSettings prims in USD files.
Note that scene-provided render settings require either
HD_PRMAN_RENDER_SETTINGS_DRIVE_RENDER_PASS=1 or
HD_PRMAN_INTERACTIVE_RENDER_WITH_RENDER_SETTINGS=1.
Add test coverage for both cases.

The pixel filter requires special handling: in the RenderMan
schemas it is a scene-global setting, similar to an Option,
but in the Riely API it needs to passed to each output.

The testHdPrman script sets a fallback pixel filter of
box 1,1 to match the RenderMan defaults prior to Rman 26,
which were also used prior to this change in the case
of scene-defined render settings (regardless of rman
version).

(Internal change: 2379578)