Fix mesh light render updates.
Previously a mesh light's xform, intensity, exposure, etc. could change
during an interactive render without triggering an update. This was
because the flattening scene index used by primvar pruning (situated
near the end of the scene index chain) wasn't evicting the relevant
cache entries so stale values were persisting. It's the responsibility
of any "setter" scene indices to ensure that prims are dirtied as
needed, so here we have the mesh light scene index use the magic
ComputeDirtyLocators() function to invalidate the locators on its
synthesized mesh+light prims that are required to initiate a cache
refresh for those prims in the flattening scene index downstream.

(Internal change: 2379606)