Teach usdview's spline viewer how to display knots.
Knots are drawn as small rectangles. If the segment type is curve then
tangents are also drawn as lines with smaller rectangles at their ends.
If the knot is dual-valued, the left half of the rectangle is drawn at
the preValue and the right half of the rectangle is drawn at the value.
The sampled version of the spline and the unrolled looped knots are
only computed when the attribute or the time range changes. The cached
spline curve is converted to a single QPainterPath object and simply
drawn. A similar optimization could possibly be done for the knots but
is not part of this change.

(Internal change: 2379628)