sdf: Replace use of tbb::parallel_for with WorkParallelForTBBRange
Going through libWork adds support for clients with custom task
management systems. This drops the use of tbb::static_partitioner since
there currently is no matching concept defined in libWork. If there is
performance fallout, we could consider adding it to the API like we
did for the RangeType concept.

(Internal change: 2379721)