Store compilation version with each compiled output.
The Exec_Program now tracks a "compilation version" which is an integer that
increases with each round of compilation. The Exec_CompiledOutputCache now
tags each output with the compilation version at the time that output was
compiled.

This is all in preparation for cycle-detection. When we find an output in
the Exec_CompiledOutputCache, we need to check if that output was added by
the current round of compilation. If not, then making the connection might
introduce a cycle.

(Internal change: 2379746)