pcp: Add PcpNodeIterator::MoveToNextSubtree
This allows clients to move between subtrees in a prim index.
For example, when iterating over the nodes in a prim index, if
a client sees a node marked as culled (which means that all
descendant nodes are also culled), they can use this API to
skip over that entire subtree and move to the next one.

(Internal change: 2379921)