usdUI: Write recently relocated hint fields to both new and old locations
The displayName, displayGroup, and hidden core fields are in the process of
moving into the uiHints dictionary, with API for the new locations provided
by UsdUIObjectHints and UsdUIPrimHints. For the time being we want this API
to write to the old locations as well, to give client code time to adapt.

This legacy write behavior is controlled by a new environment variable,
USD_WRITE_LEGACY_UI_HINT_FIELDS, which for now is true by default.

(Internal change: 2379950)