sdr, usdUtils: Populate open pages, arraySizeConstraint and limits values
SdrShaderNode now provides a list of pages that should be open/expanded by
default, and SdrShaderProperty provides array tuple-size. The latter is used
to populate arraySizeConstraint.

Min and max values are (for now) parsed manually by UpdateSchemaWithSdrNode()
and used to populate the UsdAttribute limits dictionary. In the future
SdrShaderProperty should provide proper API for this, as well as the logic
that determines which values become soft limits vs. hard limits.

Updated the args and OSL parser plugins to provide open pages.

(Internal change: 2380040)