[hdPrman] Replace usage of HdContainerDataSourceEditor in GetPrim when overriding the light prim data source.
By overlaying a lazy container data source wherein the computation is deferred and the result cached, we mimic a retained data source. This imposes additional responsibility to invalidate the prim container handle when the computed dependencies may be different.
Replace this with a lazy container data source that evaluates the dependencies on each query.
This addresses stale prim container handles in downstream scene indices that cache it (like the flattening scene index) by making it safe to use the cached prim container handle to query the dependencies data source.

A follow-up change will address the same issue for volume dependencies.

(Internal change: 2380047)