sdf: Guard against invalid primChildren in crate files
A crate file containing a primChildren field on a prim with invalid
entries (e.g. an empty name or name with invalid characters)
could lead to issues when accessing that prim's children later
on. In addition, duplicate entries in the field could cause
unexpected behavior when iterating over those children.

We now check for these conditions when loading a crate file and
emit a runtime error and fail to load the layer if we encounter
this problem.

This situation typically does not occur via regular Sdf and
Usd usage, so we only do this verification if
PXR_PREFER_SAFETY_OVER_SPEED is enabled.

(Internal change: 2380262)
(Internal change: 2380126)