pcp: Rework Pcp_Dependencies::Add to avoid computing node dependency types
Pcp_Dependencies wants to store information from every node in a prim
index that is either a direct dependency or a descendant of a direct
dependency. To do this, Add would call PcpClassifyNodeDependency on
each node to check whether it fell into this category. However,
PcpClassifyNodeDependency computed this by walking up from each node
to the root of the prim index, which meant that Add was redundantly
walking up ancestor chains.

This change simplifies Add to walk the prim index and look for direct
dependency nodes only. Once we find such a node, we know that all of
the nodes in that subtree will fall into the category above, so we
can just process that subtree without calling PcpClassifyNodeDependency
on each node.

(Internal change: 2380267)