ts: Fix various spline issues
TsSpline::Sample() was failing to generate samples for curved segments
that occurred in time-reversed oscillating extrapoation loops. It was
also failing to produce samples for sections of a curved segment if the
individual section completely spanned the sampling interval.

These failures were not initially caught because testTsSplineSampling
was not adequately validating the test results. So first we updated the
test.

The test now correctly measures the distance from the sampled polyline
to the spline (it was previously always 0). The test also verifies that
the spline is close to the sampled polyline as well as testing that the
polyline is close to the spline. This change caught several cases of
missing polyline samples that were not caught before. The test baseline
has been updated since there are now sampled points that were
previously missing in the baseline.

In addition, more carefully unrolling the spline's knots has fixed a
crash due to indexing beyond the end of the unrolledKnots array.

(Internal change: 2380311)