sdf, usd: Add deprecation warnings for migrating UI hint fields
Adds the SDF_LEGACY_UI_HINTS_WARN_ON_WRITE environment variable to enable
deprecation warnings in "set" API for the displayName, displayGroup, and
hidden metadata fields, at both the Sdf and Usd levels. The variable is set
to false by default for now but will be enabled at some point in the
future.

These fields are in the process of moving into the uiHints metadata
dictionary, which has new API available in the UsdUI library.

(Internal change: 2380364)