pcp: Maintain node restriction depth for namespace descendants
In certain cases the node restriction depth could be changed
to a deeper level of namespace than where it was originally
set, which is misleading and loses some information. This
change makes it so that the restriction depth won't be
reset in those cases.

This bug didn't appear to cause any practical issues but
did cause some unit test failures in an upcoming change so
I'm landing this as a precursor.

(Internal change: 2380374)