pcp: Simplify culling pass
Prior to simplified specializes the culling pass done at the
end of prim indexing had to take into account the duplication
of subtrees under the origin and propagated specializes arcs
to ensure that nodes were marked as culled consistently
between both trees. With simplified specializes, we no longer
have this duplication, so the culling pass can also be
simplified significantly.

This revealed an oversight in PcpPrimIndex_Graph in the code
that processed the removal of culled nodes from the graph. That
code didn't handle the case where a culled node that was marked
as being non-removable (e.g. it's the origin to another non-removal
node) had a parent node with an origin that had previously been
marked as removable. PcpPrimIndex_Graph would remove that parent's
origin, breaking the expected graph structure. This has been
fixed by recursively processing origins in this case. AFAICT
this situation only comes up with specializes arcs, which is why
we hadn't seen it before.

This change also fixes a memory regression introduced by
change 2379489. That change (correctly) caused more specializes
nodes to be composed. Some of these nodes were not culled as
expected due to the old culling code, which led to a cascade
of other nodes not being culled.

(Internal change: 2380378)