Remove verbose mip level warning trace from Hio_OpenEXRImage
Due to the code inside `HdStTextureUtils::GetAllMipImages` always trying to load 32 mip levels from each image, the warning trace inside `Hio_OpenEXRImage::_OpenForReading` is guaranteed to fire for every EXR in existence.

This is probably not necessary. Additionally, the other `Hio_*` image plugins are not nearly this verbose. For example, none of the others even trace a warning if they cannot open the image, yet the EXR plugin also does so.

Closes #3837

(Internal change: 2381155)