[hd] HdDependencyForwardingSceneIndex notice handler improvements.
Optimized _PrimsDirtied: vectorized locator checking, added a parallel for, and optimized the set of paths we check for garbage collection.  _PrimsDirtied was taking quite a bit of time in heavy scene playback, due to inefficient processing of an impressively large number of dependencies.  There are further improvements we can make to the dependency maps (noted in comments), but they are more invasive and with these improvements the dependency forwarding scene index is no longer a major contributor to frametime (in local testing).

Fixed semantics of _PrimsAdded (to correctly invalidate prims that depend on a resynced prim) and _PrimsRemoved (to interpret the removed path as a subtree instead of a prim; i.e. to correctly invalidate prims that depend on a child of a removed prim) and added a parallel for to compensate for the fact that these functions now do much more work than before.

(Internal change: 2381244)
(Internal change: 2381341)
(Internal change: 2381417)