pcp: Fix prim stack update in response to variable expression changes
If a sublayer variable expression was modified but still evaluated to
the same asset path, Pcp would incorrectly ignore that sublayer when
updating the prim stacks for affected prims.

This was because Pcp would see the sublayer modification as the removal
of the old expression followed by the addition of the new expression.
The former would add the sublayer to the layersAffectedByMutingOrRemoval
set in PcpCacheChanges; Pcp_RescanForSpecs would subsequently ignore
any sublayers in that set.

Change processing must refer to this set when performing analysis for
incremental changes becaused removed/muted sublayers are still included
in their layer stacks at that time. However, Pcp_RescanForSpecs is
called _after_ this phase, when changes to layer stacks have already
been applied. So, this function doesn't actually need to refer to the
layersAffectedByMutingOrRemoval set; removing this fixes the bug.

Ideally the affected sublayer would not be included in the
layersAffectedByMutingOrRemoval set since that sublayer wasn't
actually removed. This would also fix the issue above, but is a much
bigger change so we opt to land this intermediate fix first.

(Internal change: 2381265)