Fix exponential runtime in Vdf_DotGrapher
Traversal in Vdf_DotGrapher::_GetLimitedNodes was only constrained by depth, not whether a node has been visited. This was to ensure that each traversal would always find nodes within the requested depths (change 1830101), however it could lead to exponential runtime.

Instead, use a separate set of visited nodes for each traversal, so a traversal is still not constrained by previous traversals but will not visit a node multiple times.

Added a test which times out without the fix.

(Internal change: 2381566)