Autodesk: Fix assumption of pre-sorted render tags
Prevent crash on Windows when toggling the view rendering
representation from "smoothHull" to "hull" and then to "wireOnSurf".
This was due to HdDirtyList::UpdateRenderTagsAndReprSelectors()
previously using std::set_union() with an implicit assumption that the
provided render tags are pre-sorted.  Rather than an extra sort step,
use a more straightforward merge.

Closes #3842

(Internal change: 2381839)