pcp: Rework culled dependency processing to avoid computing node dependency types
Pcp_AddCulledDependency was previously called on each culled node
during the culling pass at the end of prim indexing. This function
called PcpClassifyNodeDependency to determine if that node should be
recorded as a dependency, which would incurred the same costs as
described in change 2380267.

To avoid this, we introduce a new Pcp_AddCulledDependencies function
that does a separate traversal of the prim index after the culling
pass has completed. This function uses the same optimization as
in the above change: it only processes nodes in subtrees beneath
direct dependencies that are discovered during the traversal.

Additionally, this change modifies Pcp_Dependencies::Add and Remove to
avoid processing culled subtrees. This was redundant work, since these
subtrees were already being handled as part of the culled dependency
processing.

(Internal change: 2382732)