HdRenderIndex::GetSceneDelegateAndInstancerIds is working when there is no _emulationSceneIndex.
This is work to be able to enable USDIMAGINGGL_ENGINE_ENABLE_SCENE_INDEX_OBSERVER_RENDERER.

The change is breaking up the if-branches on _emulationSceneIndex and _IsEnabledSceneIndexEmulation into more narrowly scoped and fine-grained if's that check for the _emulationSceneIndex and scene index adapter and _siSd.

This way, we get results for delegate id and instancer id when a HdRenderIndex is used for pure front end emulation or pure back end emulation or both.

(Internal change: 2382869)