Use UsdGeomXformCommonAPI to retrieve the pivot primvar.
It isn't sufficient in all cases to simply ask for the value of the
xformOp:translate:pivot USD attribute. Sometimes a prim might not have
that attribute but it will have a rotate and/or scale pivot, e.g.
xformOp:translate:rotatePivot, which still counts as a pivot translation
in UsdGeomXformCommonAPI (where this pivot concept originates from). In
such cases we would want to use those other xformOp values as the pivot,
but the current code can only look for an xformOp attribute with the
exact name "pivot". It may even use an inherited pivot if it can't find
one on the prim itself.

UsdGeomXformCommonAPI::GetXformVectorsByAccumulation() already
implements the pivot-getting behavior that we want, so it seems simplest
to use that in the primvars data source when something asks for the
special-case pivot primvar.

We've updated the name of the pivot primvar to xformOp:computedPivot
because we're no longer just reading xformOp:translate:pivot directly.

(Internal change: 2382965)