[hdSt] Sort textures by size before committing to GL.
Credit to Sunya: while investigating variance in perf test runtimes, he noticed that if we sort textures prior to commit we both speed up texture commit generally and also cut the variance way down.

This change adds a "GetCommittedSize" to HdStTextureObject, which is intended to be the number of bytes (cumulatively, for ptex/udim) that are passed to CreateTexture.  We sort biggest-to-smallest, as in preliminary testing this sort order was faster than smallest-to-biggest.

While this sort order isn't perfectly stable run to run (it's still possible for two textures of the same size to swap back and forth), the hope is that this provides a more consistent runtime for texture upload.

(Internal change: 2384003)