[sdf] SdfPrimSpec::New() no longer substitutes __AnyType__ for an empty typename.
__AnyType__ was a workaround for the early file parser's
inability to allow empty typenames in defs, circa 2010.
Empty typenames in defs have been well-supported since 2013,
so we no longer need to do this, and by omitting it we save
a little bit of extra noise in files.  Note that __AnyType__
continues to be allowed and supported; we simply no longer
insert that in place of an empty typename.

The typical scenario where a typeless def occurs is when
the typename is provided across a reference arc.

(Internal change: 2384346)