Implement a TsSpline iterator.
The spline iterator is implemented in 2 primary classes, Ts_Segment and
Ts_SegmentIterator. There are also 3 helper iterators, Ts_SegmentKnotIterator,
Ts_SegmentLoopIterator, and Ts_SegmentPrototypeIterator.

Ts_Segment holds a segment of a spline. A segment is the span between two knots.
The segment contains the endpoints of the span, tangent information, and an
interpolation type for the span. There are also special pre- and
post-extrapolation interpolation types that extend to infinity with a slope
rather than an end point (since an end point at infinity is problematic).

A Ts_SegmentIterator is constructed from a spline and a time interval to iterate
over. Dereferencing the iterator returns a Ts_Segment. Because splines are
effectively infinite, the iterator's AtEnd() method returns true when the end of
the interval is reached.

Under the hood, the Ts_SegmentIterator uses a Ts_KnotIterator to iterate over
just the knots in the spline, possibly repeatedly when there is extrapolation
looping in effect.

Ts_SegmentKnotIterator uses Ts_SegmentLoopIterator to iterate over any inner
looped section, and Ts_SegmentLoopIterator uses Ts_SegmentPrototypeIterator as
needed to get the prototype segments.

(Internal change: 2384416)