[MaterialX] Fix edge cases
This change fixes two edge cases.
The first is we were not properly getting the nodedef for locally defined custom multi-output nodes.
The second was the case that custom nodes parameter inputs have leading underscores. MaterialX concatenates most node inputs with their node name and an underscore as the separator. During this concatenation it removes the leading underscore beforehand. We were not doing that so we were not getting any authored values or textures because of the name mismatch.

Added test cases for leading underscores on texture and non-texture parameter inputs for surfaceshader and non-surfaceshader custom nodes in both the library and locally defined tests, in usdImagingGL and HdPrman.

(Internal change: 2384525)