[usdRi] Add utilities for wokring with RenderMan attributes that use its set specification algebra syntax.
The common example of this is a grouping:membership attribute
value with a + operator, such as "+Foo", indicating that
geometry should be added to the set "Foo" while preserving
membership in other sets.  This algebra is resolved when
transferring attributes from materials to geometry that binds
the material.

Since the underlying RenderMan code to work with set
membership in string form is not exposed, we add some
utilities in UsdRi that can convert this syntax to
a SdfStringListOp.  From there, the standard Sdf
API can be used to combine this with other lists.

Also introduce UsdRiDoesAttributeUseSetSpecification(name)
to provide a way to centralize knowledge of which attributes
use these semantics -- since it is not represented in the
type system or RenderMan schema metadata.

(Internal change: 2384682)