pcp: Add test exercising subroot referencing and instancing behavior
Currently, when a child prim of an instance is referenced by another
prim, any local opinions about that child prim are ignored. This
wasn't intentionally designed and there were questions about
whether subroot references to a prim beneath an instance ought to
ignore the ancestral "instanceable" tag and allow the opinions
to come through instead.

We ultimately decided on leaving the current behavior in place.
We felt that a user would find it surprising if they saw a different
set of opinions when referencing a prim than what they'd see if
they looked at the referenced prim directly. This change just
codifies this behavior in a museum test case.

(Internal change: 2385039)