[hdSt] Revert changelist 2384003 (sort textures before upload).
This change was an attempt to optimize texture upload (and reduce variance of upload time) inside the GL driver by sorting textures by size before submission.  This showed promise in local & farm testing, but doesn't appear to have moved the needle in continuous integration.  It also appears to have regressed behavior in the OpenUSD build when using normal malloc (as opposed to jemalloc).  Alas...

(Internal change: 2385353)