[hd] HdMaterialSchema API tweaks informed by user feedback
GetMaterialNetwork(renderContext) now retrieves the material
for the provided render context only, and does not implicitly
fall back to the universal context.

The prior behavior of having an implicit fallback to
the universal context was found to be confusing by
authors of scene indexes that want to process material
networks for specific render contexts.

Make these read-only schema methods const.

(Internal change: 2385805)