pcp: Ignore nodes that cannot contribute specs in instancing key
When propagating implied inherit or specializes nodes in a prim
index, there are cases where the implied node winds up with the
same site as its origin node. Those nodes are left in the prim
index but marked as inert to indicate that they do not contribute
opinions but are there as placeholders for the algorithm to
continue.

Since these inert nodes can't contribute opinions, they should
not be considered part of the instancing key for a prim index.
This fixes cases where UsdStage wasn't optimally sharing
prototypes between instances -- see updated test cases for
simple examples.

(Internal change: 2386369)