pcp: Add self-consistency check for diagnostics and testing
Pcp will now perform self-consistency checks on all composed
prim indexes when the env var PCP_ENABLE_CONSISTENCY_CHECK
is set to 1. This provides a way to verify the structure
of prim indexes outside of specifically-written unit tests,
making it easier to test a wide variety of assets. These
checks are potentially expensive and are intended to be used
for testing and debugging.

The self-consistency check currently just verifies that the
"culling" flag is set correctly on nodes, but other checks
will follow. This check is enabled for all unit tests in Pcp.

(Internal change: 2386423)