vt: Support pieces for new USD value resolution framework.
- Add VtValueRef, a non-owning VtValue-like view on a value type, mostly
useful for function arguments.  Both ordinary objects and VtValue can
implicitly convert to VtValueRef.

- Add general VtValue compose-over functionality.  This lets value types
that can compose over other value types (e.g. VtDictionary, VtArrayEdit,
SdfPathExpression, SdfListOp) register with VtValue, and then these
values can be composed generically with Vt API.  Moreover, compositions
of composing types work automatically for Vt types, and can be extended
for third-party types.  For example, since VtArrayEdit composes over
other VtArrayEdits and VtArrays, and SdfPathExpressions compose over
themselves, now a VtDictionary containing an array edit in a stronger
dictionary's key will compose over an array or an array edit at the same
key in a weaker dictionary.  Path expressions or SdfListOps contained in
dictionaries will behave similarly.  There is a special VtBackground
object that can be used to "finalize" composing values.  For example,
when a VtArrayEdit is composed over the VtBackground, it applies its
edits to an empty array to produce an array result.  Composing
SdfListOps over the VtBackground will similarly apply their edits to an
empty item vector and yield an SdfListOp with those explicit items.
Composing a VtDictionary over the VtBackground just composes all of its
values over the VtBackground.  To be sensible, compose-over functions
must be associative.

- Add general VtValue transform functionality.  Value transforms produce
a new value of the same type as the input, and need not be associative
like compose-over, but must yield an object of the same type.  Similarly
to compose-over functionality, registered transforms automatically apply
to values in VtDictionary, and VtArray and VtArrayEdit for array-
supported types.  In practice we will use this in USD to provide a
single straightforward way to translate values from layers into the name-
and time-space of a UsdStage, and the inverse.  This includes path-
translation, layer-offset time mapping, and prototype-to-instance
mappings.

(Internal change: 2386428)