Fix calculation of the slope of TsSpline linear extrapolation
If the last knot before extrapolation was generated by inner looping,
the calculation of the preValue of that knot was incorrect. This caused
an incorrect slope to be computed. The test failed to detect this
because the correct slope was 0.0 and the erroneously computed slope was
also 0.0.

The test has been updated to exercise this case by ensuring that the
correct slope is non-zero.

(Internal change: 2386654)