[usdUtils] Small additive API to specify a path for variant selections
Adds a new overload for `GetSessionLayerForVariantSelections` that accepts a path where variant selections should be.

There is a slight behavior change in that the sessionLayer key will contain a leading `/` now.

This also does some minor cleanup of removing `TF_FOR_ALL` and adds argument names to the python wrapping.

(Internal change: 2386994)