Implement TsSpline::Diff() to compare two splines.
Diff returns a GfInterval which is the time interval over which the two
splines differ. There are 2 overloaded methods, the first simply
compares the splines across all time, -inf to +inf. The second takes a
compareInterval argument and returns differences that are only in the
compareInterval.

There are a couple of caveats. Diff compares the "segments" of the
spline where a segment is the span between adjacent knots. If the
segments have different interpolation then they are different. So a
held segment may compare as different than a linear segment, even if
the linear segment represents the same horizontal span. Similarly, if
there is a knot that does not change the shape of the spline it will
still change the size of the segments and that section will be reported
as different.

Finally, extrapolation has its own special issues. All extrapolation is
either a value-block, a straight geometric ray from an endpoint off to
infinity, or some repeated loop over the spline's knots. Value-blocks
with different endpoints currently report they are different to
infinity. Looping extrapolation of different types is also assumed to
be different to infinity. But rays that are identical are reported to
be the same, even if they come from different extrapolation types (held
vs sloped for example).

(Internal change: 2387251)