Add TestExecUsdRecompilation_Perf
This adds a series of tests that measure the performance of recompilation in
OpenExec. Each test alters relationship targets in some way, which causes
the uncompilation and subsequent recompilation of individual inputs without
uncompiling any nodes.

These tests are being added now so we have metrics in place for evaluating
different implementations of cycle detection.

(Internal change: 2387563)