build: Add RENDERMAN_OSL_INCLUDE_DIR
Add RENDERMAN_OSL_INCLUDE_DIR to FindRenderman.cmake and use it when
building OSL shaders.

When building the renderman plugin, often the OSL include to get
`stdosl.h` is left undefined and will try to find it either from
another OSL build and not the one found bundled with Renderman, or in
Windows' case, will find one with forward-slashes or a resolved UNC
path from a network and fail to find the file.

Closes #3799

(Internal change: 2387915)