usdImaging: Avoid use of inf as a sentinel time in UsdImagingDelegate
In certain cases this would lead to a UsdAttribute being queried at
UsdTimeCode(inf), which "works" when fetching values from time samples
but causes failed verifies with value clips.

Creating UsdTimeCode with inf will become a coding error in a future
change; to accommodate this UsdImagingDelegate now uses the max double
value as its sentinel instead. This should have the same behavior as
before with time samples, but avoid the issue with value clips.

Although UsdImagingDelegate is being phased out for Hydra 2.0, we
still need to update this code if for no other reason than to avoid
breaking existing unit tests with the future UsdTimeCode change.

(Internal change: 2388071)