Pass additional color space authoring methods into Hydra
For UsdImaging: The DataSourceAttributeColorspace now uses
UsdColorSpaceAPI::ComputeColorSpaceName() instead of
attr.GetColorSpace() to get the authored color space as it resolves
color spaces authored as attribute metadata and through the
UsdColorSpaceAPI.

This change also adds some special case handling for UsdUVTexture Nodes.
As that node has a separate 'sourceColorSpace' input. This change will
consolidate that authored input value, as the colorSpace on the file
attribute instead of its own separate attribute. This means that to Hydra
the color space will always be formatted the same.

Corresponding updates have been made in hdSt/hdPrman
to get the color space information from this one new location.

(Internal change: 2388446)