sdf: Fix target/connection performance issue in SdfLayer::CreateDiff
Replace calls to SdfAbstractData::GetSpecType with a helper that
derives the spec type for target paths based on the spec type
of its owning property. This avoids a performance issue due to
Usd_CrateData having to synthesize this information from the
target/connection list op.

This showed up when adding or removing .usdc-based sublayers
with relationships with many targets or attributes with many
connections to a UsdStage. The original GitHub issue filed about
this had a sublayer containing a single relationship with 10000
targets. Prior to this change, adding that sublayer took about
20 seconds. After this change, this time drops to 0.1 seconds,
which is the same as if the sublayer was .usda-based instead.

Fixes #3876

(Internal change: 2389267)